requirements.txt
absl-py==2.1.0
einops==0.7.0
einsum==0.3.0
huggingface-hub==0.20.3
ml-collections==0.1.1
numpy==1.26.2
pandas==2.2.0
scikit-learn==1.4.1.post1
scipy==1.12.0
tensorflow>=2.12.1
timm==0.9.16
torch==2.1.2
torchaudio==2.1.2
torchdata==0.7.1
torchmetrics==1.3.1
torchvision==0.16.2
transformers==4.37.2
triton==2.1.0
